Refactor: 코드리뷰 반영
컴포넌트 클래스 이름 변경, 모델 프로퍼티 이름 변경
