--- MUMENT/MUMENT/Sources/Components/MumentWithHeartView.swift
@@ -1,69 +1,0 @@
-//
-//  MumentWithHeart.swift
-//  MUMENT
-//
-//  Created by 김지민 on 2022/07/13.
-//
-
-import UIKit
-import SnapKit
-import Then
-
-class MumentWithHeartView: DefaultMumentView {
-
-    // MARK: - Properties
-    private let heartButton = UIButton().then{
-        var configuration = UIButton.Configuration.plain()
-            configuration.imagePadding = 5
-            configuration.buttonSize = .small
-        $0.configuration = configuration
-    }
-    
-    let attributes: [NSAttributedString.Key: Any] = [
-        .font: UIFont.mumentC1R12,
-        .foregroundColor: UIColor.mGray1
-    ]
-    
-    var heartButtonText: String = "" {
-        didSet{
-            heartButton.setAttributedTitle(NSAttributedString(string: heartButtonText,attributes: attributes), for: .normal)
-        }
-    }
-    
-    // MARK: - Initialization
-    override init(frame: CGRect) {
-        super.init(frame: .zero)
-        setLayout()
-    }
-    
-    required init(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-    }
-    
-    //MARK: - Functions
-    func setData(_ cellData: MumentWithHeartModel){
-        profileImage.image = cellData.profileImage
-        writerNameLabel.text = cellData.writerName
-        albumImage.image = cellData.albumImage
-        songTitleLabel.text = cellData.songTitle
-        artistLabel.text = cellData.artistName
-        contentsLabel.text = cellData.contentsLabel
-        createdAtLabel.text = cellData.createdAtLabel
-        heartButton.setImage(cellData.heartImage, for: .normal)
-        heartButtonText = "\(cellData.heartCount)"
-    }
-}
-
-// MARK: - UI
-extension MumentWithHeartView {
-    func setLayout() {
-        self.addSubview(heartButton)
-
-        heartButton.snp.makeConstraints {
-            $0.right.equalTo(self.safeAreaLayoutGuide).inset(5)
-            $0.top.equalTo(self.safeAreaLayoutGuide).offset(9)
-        }
-
-    }
-}
-
