+++ MUMENT/MUMENT/Sources/Scenes/Home/MumentForTodayTVC.swift
@@ -12,13 +12,13 @@ import Then
 class MumentForTodayTVC: UITableViewCell {
     
     // MARK: - Properties
-    var dataSource: [MumentWithHeartModel] = MumentWithHeartModel.sampleData
+    var dataSource: [DefaultMumentModel] = DefaultMumentModel.sampleData
     lazy var titleLabel = UILabel().then{
         $0.text = "오늘의 뮤멘트"
         $0.textColor = .mBlack1
         $0.font = .mumentH2B18
     }
-    private let mumentCardView = MumentWithHeartView()
+    private let mumentCardView = DefaultMumentView()
     
     // MARK: - Initialization
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
